Actualizacion grande
* Repositorio actualizado
* Carpetas de Debug actualizadas
* Problema con ejecutables!
* Lecturas de TODAS las Practicas actualizadas
* Nuevos comentarios y cuerpos de funciones creados
* Pendiente arreglos

+++ PI3FernandezTrigoAlejandro/ejercicios/ejercicio3iterativa.h
@@ -16,6 +16,6 @@
 #include "types/list.h"
 
 // Cabeceras de las funciones:
-int ejercicio3Iterativo (int numero);
+long ejercicio3Iterativo (int numero);
 
 #endif /* EJERCICIOS_EJERCICIO3ITERATIVA_H_ */
